+++ spec/injectable_spec.rb
@@ -162,8 +162,8 @@ describe Injectable do
   context 'with dependencies that have a call: option' do
     before do
       SomeRenderer = Class.new do
-        def render
-          '#render has been called'
+        def render(arg, kwarg:)
+          "#render has been called with #{arg} and #{kwarg}"
         end
       end
     end
@@ -173,12 +173,15 @@ describe Injectable do
         include Injectable
         extend Forwardable
         dependency :some_renderer, call: :render
-        def_delegators :some_renderer, :call
+
+        def call
+          some_renderer.call('hello', kwarg: 'world')
+        end
       end
     end
 
     it 'wraps the specified method in a #call method' do
-      expect(subject.call).to eq '#render has been called'
+      expect(subject.call).to eq '#render has been called with hello and world'
     end
   end
 
@@ -491,6 +494,58 @@ describe Injectable do
 
     subject { [Parent.call, Child.call, Sibling.call] }
 
-    it { is_expected.to eq ['in parent', 'in child', 'in sibling']}
+    it { is_expected.to eq ['in parent', 'in child', 'in sibling'] }
+  end
+
+  context 'when the dependency accepts a block' do
+    before do
+      class BlockPasser
+        def call
+          yield
+        end
+      end
+    end
+
+    subject do
+      Class.new do
+        include Injectable
+
+        dependency :block_passer
+
+        def call
+          block_passer.call { "can't block this" }
+        end
+      end
+    end
+
+    it 'passes the block to the dependency' do
+      expect(subject.call).to eq "can't block this"
+    end
+  end
+
+  context 'when the dependency accepts a block and has #call aliased' do
+    before do
+      class BlockPasser
+        def run
+          yield
+        end
+      end
+    end
+
+    subject do
+      Class.new do
+        include Injectable
+
+        dependency :block_passer, call: :run
+
+        def call
+          block_passer.call { "can't block this" }
+        end
+      end
+    end
+
+    it 'passes the block to the dependency' do
+      expect(subject.call).to eq "can't block this"
+    end
   end
 end