+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Fixed a bug that wouldn't pass a block if `#call` was aliased (#7)
+
 ### 1.0.2 - 2020-03-02
 
 * Bump `rake` to 12.3.3 (#11)